[refactor] net-perf/src/data/iperf3.json: 프리티어 적용
